@@ -1,11 +1,10 @@
 module github.com/mataharibiz/ward
 
-go 1.20
+go 1.23
 
 require (
 	github.com/google/uuid v1.6.0
 	github.com/rabbitmq/amqp091-go v1.10.0
-	github.com/sirupsen/logrus v1.9.3
 )
 
 require (
