@@ -1,14 +1,5 @@
 package logging
 
-import "github.com/sirupsen/logrus"
-
-func (log *LogEntry) Warn(data interface{}, message string) {
-	if data != nil {
-		log.logger.WithFields(logrus.Fields{
-			"data": data,
-		}).Warn(message)
-		return
-	}
-
-	log.logger.Warn(message)
+func (log *LogEntry) Warn(message string, args ...any) {
+	log.logger.Warn(message, log.appLogger(args...)...)
 }
